move root layout.tsx and page.tsx in src/app/[locale] + add next-intl config
