--- src/app/layout.tsx
@@ -1,31 +1,0 @@
-// types
-import type { Metadata } from "next";
-
-// styles
-import "./globals.css";
-
-// themes provider/switchers
-import ThemeContextProvider from "@/context/theme-provider";
-import ThemeSwitcher from "@/components/theme-switcher";
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className="bg-amber-50 antialiased dark:bg-zinc-800">
-        <ThemeContextProvider>
-          <ThemeSwitcher />
-          <main>{children}</main>
-        </ThemeContextProvider>
-      </body>
-    </html>
-  );
-}
